collection sequence fix (#3)
We were dropping `CollectionSequenceNumber` when running through the
serialize (in nest we call in normalize) step. Looks like there are no
other `element.*:string` definitions, and other similar/simple elements
use `:text`.

I will admit that I didn't trace down exactly where this was causing
issues, and/or why something didn't fail a bit louder when an invalid
type was used.

Either way, this is working in nest, and there are no other
`element.*:string` to update, so this looks like an isolated bug.